Refactor help to MVC
